celery did not follow up on their promise, and it's my problem somehow
