+++ env.example
@@ -12,5 +12,4 @@ POSTGRES_HOST=#DB host
 POSTGRES_PORT=5432
 TELEGRAM_KEY=#Telegram Bot API key 
 REMINDER_CHAT_IDS=#Telegram chat IDs, to which the reminders will be sent
-MQ_USER=#RabbitMQ username
-MQ_PASSWORD=#RabbitMQ user password
+MQ_BROKER_URL=#RabbitMQ broker URL
